+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Ignore sub-pixel height changes by rounding before size-change callback
+
 ## [0.5.1] - 2025-06-18
 
 ### Added