+++ CHANGELOG.md
@@ -11,7 +11,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add support to all distributions for clickable progress
 - Fallback `imageKey` support in `DataImage` and `DataImageCarousel` nodes
+
+### Fixed
+
+- Ignore sub-pixel height changes by rounding before size-change callback
 
 ## [0.5.1] - 2025-06-18